+++ packages/orq-rc/src/models/operations/fileget.ts
@@ -94,7 +94,7 @@ export const FileGetResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-23T08:50:22.186Z",
+    "2026-02-23T10:17:38.967Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {