+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-23T08:50:18.831Z",
+    "2026-02-23T10:17:35.403Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {