+++ packages/orq-rc/RELEASES.md
@@ -7109,3 +7109,13 @@ Based on:
 - [typescript v4.4.0-rc.28] packages/orq-rc
 ### Releases
 - [NPM v4.4.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.28 - packages/orq-rc
+
+## 2026-02-23 10:20:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.724.1 (2.836.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.4.0-rc.29] packages/orq-rc
+### Releases
+- [NPM v4.4.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.29 - packages/orq-rc